Cast avoid signed warning
git-svn-id: https://svn.blender.org/svnroot/verse/branches/release-r6-branch@3727 49c17534-1d02-dc11-b283-0030488c597c

+++ vs_main.c
@@ -113,7 +113,7 @@ int main(int argc, char **argv)
 
 	signal(SIGINT, cb_sigint_handler);
 
-	for(i = 1; i < argc; i++)
+	for(i = 1; i < (uint32) argc; i++)
 	{
 		if(strcmp(argv[i], "-Q") == 0)
 			ms_address = NULL;
